Added "uid" and "pidfile" options.
"uid" allows you to drop priveleges after daemonizing. Its argument can
either be a numeric UID or a username that can be found in /etc/passwd.

"pidfile" is the name of a file to write the daemon's PID to. For daemons
that fork, none of the children will have their PIDs added to the file.